+++ libs/import/src/lib/import.component.ts
@@ -1,15 +1,15 @@
 import { Component, ViewChild, ChangeDetectionStrategy, ChangeDetectorRef, OnInit } from '@angular/core';
 import { DynamicTitleService } from '@blockframes/utils/dynamic-title/dynamic-title.service';
 import { AuthQuery } from '@blockframes/auth/+state';
-import { ViewExtractedMoviesComponent } from '../../components/view-extracted-elements/movies/view-extracted-movies.component';
-import { ViewExtractedContractsComponent } from '../../components/view-extracted-elements/contract/view-extracted-contracts.component';
-import { ViewExtractedOrganizationsComponent } from '../../components/view-extracted-elements/organizations/view-extracted-organizations.component';
-import { SpreadsheetImportEvent } from '../../components/import-spreadsheet/import-spreadsheet.component';
+import { ViewExtractedMoviesComponent } from './view-extracted-elements/movies/view-extracted-movies.component';
+import { ViewExtractedContractsComponent } from './view-extracted-elements/contract/view-extracted-contracts.component';
+import { ViewExtractedOrganizationsComponent } from './view-extracted-elements/organizations/view-extracted-organizations.component';
+import { SpreadsheetImportEvent } from './import-spreadsheet/import-spreadsheet.component';
 
 @Component({
   selector: 'import-container',
-  templateUrl: './import-container.component.html',
-  styleUrls: ['./import-container.component.scss'],
+  templateUrl: './import.component.html',
+  styleUrls: ['./import.component.scss'],
   changeDetection: ChangeDetectionStrategy.OnPush
 })
 export class ImportContainerComponent implements OnInit {
