+++ libs/import/src/lib/table-extracted-elements/contracts/contracts.component.scss
