+++ libs/import/src/lib/table-extracted-elements/movies/movies.component.scss
