+++ libs/import/src/lib/table-extracted-elements/organizations/organizations.component.scss
