+++ libs/import/src/lib/view-extracted-elements/contract/view-extracted-contracts.component.scss
