+++ libs/import/src/lib/view-extracted-elements/movies/view-extracted-movies.component.scss
