+++ libs/import/src/lib/view-extracted-elements/organizations/view-extracted-organizations.component.scss
