+++ libs/utils/src/lib/spreadsheet/format.ts
@@ -8,7 +8,7 @@ import {
   Movie,
   populateMovieLanguageSpecification
 } from "@blockframes/movie/+state/movie.model";
-import { MovieImportState } from "libs/import/src/lib/import-utils";
+import { MovieImportState } from "libs/import/src/lib/utils";
 import { createCredit, createStakeholder } from "../common-interfaces/identity";
 import { getKeyIfExists } from "../helpers";
 import { Scope } from "../static-model/static-model";
